+++ metadata/elastic/elasticupdate.py
@@ -1,32 +1,33 @@
 #!/usr/bin/python
 """Core interface for the uploader metadata objects to interface with CherryPy."""
-from os import getenv
 from cherrypy import tools, request
 from elasticsearch import Elasticsearch, helpers
+from metadata.elastic import ELASTIC_ENDPOINT, ELASTIC_INDEX
 from metadata.orm.base import db_connection_decorator
 from metadata.rest.objectinfo import ObjectInfoAPI
-
-ELASTIC_INDEX = getenv('ELASTIC_INDEX', 'pacifica')
-DEFAULT_ELASTIC_ENDPOINT = getenv('ELASTICDB_PORT', 'tcp://127.0.0.1:9200').replace('tcp', 'http')
-ELASTIC_ENDPOINT = getenv('ELASTIC_ENDPOINT', DEFAULT_ELASTIC_ENDPOINT)
-ES_INDEX_URL = '{0}/{1}'.format(ELASTIC_ENDPOINT, ELASTIC_INDEX)
 
 
 # pylint: disable=too-few-public-methods
 class ElasticSearchUpdateAPI(object):
     """ElasticSearchUpdateAPI."""
 
+    es_kwargs = {
+        'sniff_on_start': True,
+        'sniff_on_connection_fail': True,
+        'sniffer_timeout': 60,
+        'timeout': 60
+    }
     exposed = True
 
-    @staticmethod
+    @classmethod
     @db_connection_decorator
-    def push_elastic_updates(object_class_name, id_list, recursion_depth):
+    def push_elastic_updates(cls, object_class_name, id_list, recursion_depth):
         """Push out updates to the ES cluster for updated MD records."""
         myclass = ObjectInfoAPI.get_class_object_from_name(object_class_name=object_class_name)
         records = myclass.select()
         if id_list:
             records.where(myclass.id << id_list)
-        es_client = Elasticsearch([ELASTIC_ENDPOINT])
+        es_client = Elasticsearch([ELASTIC_ENDPOINT], **cls.es_kwargs)
         update_list = ({
             '_op_type': 'update',
             '_index': ELASTIC_INDEX,
