+++ metadata/rest/root.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python
 """CherryPy root object class."""
 from metadata.orm import ORM_OBJECTS
-from metadata.rest.elasticupdate import ElasticSearchUpdateAPI
+from metadata.elastic.elasticupdate import ElasticSearchUpdateAPI
 from metadata.rest.objectinfo import ObjectInfoAPI
 from metadata.rest.userinfo import UserInfoAPI
 from metadata.rest.proposalinfo import ProposalInfoAPI
